+++ src/main/java/org/cookieandkakao/babting/domain/meeting/exception/meeting/MeetingNotFoundException.java
@@ -1,4 +1,4 @@
-package org.cookieandkakao.babting.domain.meeting.exception;
+package org.cookieandkakao.babting.domain.meeting.exception.meeting;
 
 import java.util.NoSuchElementException;
