+++ src/main/java/org/cookieandkakao/babting/domain/meeting/exception/membermeeting/MemberMeetingNotFoundException.java
@@ -1,4 +1,4 @@
-package org.cookieandkakao.babting.domain.meeting.exception;
+package org.cookieandkakao.babting.domain.meeting.exception.membermeeting;
 
 import java.util.NoSuchElementException;
